+++ src/zeython/cli/templates/starter/app/Providers/app_event_service_provider.py
@@ -2,6 +2,7 @@ from zeython import EventDispatcher, EventServiceProvider
 
 from app.Events.user_registered import UserRegistered
 from app.Listeners.log_user_registration import log_user_registration
+from app.Listeners.notify_webhooks_of_registration import notify_webhooks_of_registration
 
 
 class AppEventServiceProvider(EventServiceProvider):
@@ -15,3 +16,4 @@ class AppEventServiceProvider(EventServiceProvider):
         super().boot()
         dispatcher = self.container.make(EventDispatcher)
         dispatcher.listen(UserRegistered, log_user_registration)
+        dispatcher.listen(UserRegistered, notify_webhooks_of_registration)